+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/ahorro/AhorroRequestNew.java
@@ -36,7 +36,7 @@ public class AhorroRequestNew implements Serializable {
     @ApiModelProperty(value = "Id de la Entidad Financiera", required = true)
     @JsonProperty("entidad_financiera_id")
     @SerializedName("entidad_financiera_id")
-    private Integer entidadFinancieraId;
+    private Long entidadFinancieraId;
 
     @ApiModelProperty(value = "Estado del Ahorro", required = true)
     @JsonProperty("estado")
